+++ examples/client/src/main/java/com/clickhouse/examples/jdbc/Main.java
@@ -94,7 +94,6 @@ public class Main {
         ClickHouseNode server = ClickHouseNode.builder()
                 .host(System.getProperty("chHost", "localhost"))
                 .port(ClickHouseProtocol.HTTP, Integer.getInteger("chPort", 8123))
-                // .port(ClickHouseProtocol.GRPC, Integer.getInteger("chPort", 9000))
                 // .port(ClickHouseProtocol.TCP, Integer.getInteger("chPort", 9100))
                 .database("system").credentials(ClickHouseCredentials.fromUserAndPassword(
                         System.getProperty("chUser", "default"), System.getProperty("chPassword", "")))